requirements.txt
certifi==2020.4.5.2
chardet==3.0.4
idna==2.9
requests==2.24.0
urllib3==1.25.9
ynab-sdk-python==0.0.1
